Refactor: Move get_case_location
